+++ src/shared/hooks/image/useUploadImagesGraphQL.ts
@@ -1,0 +1,61 @@
+import { gql, useMutation } from "@apollo/client";
+import { useCallback, useState } from "react";
+import type {
+  IMutation,
+  IMutationUploadFileArgs,
+} from "@/api/graphql/generated/types";
+
+export const UPLOAD_FILE = gql`
+  mutation uploadFile($file: Upload!) {
+    uploadFile(file: $file) {
+      _id
+      url
+    }
+  }
+`;
+
+export function useUploadImagesGraphQL() {
+  const [isUploading, setIsUploading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  const [uploadFile] = useMutation<
+    Pick<IMutation, "uploadFile">,
+    IMutationUploadFileArgs
+  >(UPLOAD_FILE);
+
+  const reset = useCallback(() => setError(null), []);
+
+  const uploadImages = useCallback(
+    async (filesToUpload?: File[]): Promise<string[]> => {
+      const files = filesToUpload ?? [];
+      if (files.length === 0) return [];
+
+      setIsUploading(true);
+      setError(null);
+
+      try {
+        const res = await Promise.all(
+          files.map((file) => uploadFile({ variables: { file } })),
+        );
+
+        return res
+          .map((r) => r.data?.uploadFile.url)
+          .filter((url): url is string => Boolean(url));
+      } catch (e) {
+        const message = e instanceof Error ? e.message : "업로드 실패";
+        setError(message);
+        throw new Error(message);
+      } finally {
+        setIsUploading(false);
+      }
+    },
+    [uploadFile],
+  );
+
+  return {
+    isUploading,
+    error,
+    uploadImages,
+    reset,
+  };
+}
